docs: add a TODO to add more failure message in the future
A user once reported that the Microsoft Addon API returned some weird
respond (#7). This is not a bug of this action, but we had better tell
user that this is something happening on the Microsoft server side. Show
an URL to redirect the user to Microsoft's API doc; it explains more.